requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
drf-nested-routers==0.94.1